test: split off common_tests from common
